Geodata builder (#2)

+++ rendering/src/VertexBuffer.cpp
@@ -22,13 +22,13 @@ auto VertexBuffer::int_layouts() const
 void VertexBuffer::float_layout(unsigned int index, std::size_t size,
                                 std::size_t offset, unsigned int divisor) {
 
-  const auto count = static_cast<int>(size / sizeof(float));
+  const auto count = size / sizeof(float);
   ASSERT(count > 0 && count <= 4, "Rendering",
          "Number of attribute components must be > 0 and <= 4");
 
   m_float_layouts.push_back({
       index,
-      count,
+      static_cast<int>(count),
       GL_FLOAT,
       GL_FALSE,
       m_vertex_size,
@@ -40,13 +40,13 @@ void VertexBuffer::float_layout(unsigned int index, std::size_t size,
 void VertexBuffer::int_layout(unsigned int index, std::size_t size,
                               std::size_t offset, unsigned int divisor) {
 
-  const auto count = static_cast<int>(size / sizeof(int));
+  const auto count = size / sizeof(int);
   ASSERT(count > 0 && count <= 4, "Rendering",
          "Number of attribute components must be > 0 and <= 4");
 
   m_int_layouts.push_back({
       index,
-      count,
+      static_cast<int>(count),
       GL_INT,
       m_vertex_size,
       reinterpret_cast<const void *>(offset),
